requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Data Handling
pandas>=1.3.0
scikit-learn>=1.0.0
pillow>=8.0.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0

# System Monitoring
psutil>=5.8.0

# Optional: Text Processing (uncomment if needed)
# transformers>=4.20.0
# tokenizers>=0.12.0

# Development and Testing
pytest>=6.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Progress Tracking
tqdm>=4.60.0